--- a/export.go
+++ b/export.go
@@ -15,7 +15,7 @@ const (
 
 // Dump dumps specified data.
 func Dump(data any, opts ...OptionFunc) string {
-	return newDataDumper(data, true, opts...).build().String()
+	return newDataDumper(data, opts...).build().String()
 }
 
 // Writer is a writer for dump string.